requirements.txt
spacy >= 3.7.6
transformers == 4.46.1
psycopg2-binary == 2.9.9
openai == 1.45.0
scikit-learn == 1.5.2
torch == 2.5.1
setuptools == 75.2.0
jellyfish == 1.1.0
pandas == 2.2.2
nltk == 3.9.1
accelerate >= 0.26.0